+++ src/number/digitize/test.ts
@@ -1,7 +1,7 @@
 import "./index";
 
 describe("Number.prototype.digitize", () => {
-  test("Number.digitize(123) returns [1, 2, 3]", () => {
-    expect(Number.digitize(123)).toEqual([1, 2, 3]);
+  test("(123).digitize() returns [1, 2, 3]", () => {
+    expect((123).digitize()).toEqual([1, 2, 3]);
   });
 });