+++ src/number/inRange/index.ts
@@ -1,19 +1,20 @@
 import * as method from "./method";
 
 declare global {
-  interface NumberConstructor {
-    inRange(num: number, end: number, start?: number): boolean;
+  interface Number {
+    inRange(end: number, start?: number): boolean;
   }
 }
 
 /**
  * Checks if n is between start and up to end
  * @memberof Number
- * @param {Number} num
  * @param {Number} end
  * @param {Number} [start=0]
  * @returns {Boolean}
  * @example
- * Number.inRange(4, 8); // true
+ * (4).inRange(8); // true
  */
-Number.inRange = method;
+Number.prototype.inRange = function(end, start) {
+  return method(this as number, end, start);
+};